Rewrite output verification tests to use unittest
Significantly simplify the verifyOutput module so that it
makes use of modern features in python and uses assertListEqual.
The driver scripts now properly use unittest and can be run
with py.test.

The test output must still be canonicalized to account for field
attributes associated with messages appearing in random order.